synchronized playback and progress bar animations when resuming play after pausing

+++ src/com/logan/feelchakra/Track.scala
@@ -1,5 +1,5 @@
 package com.logan.feelchakra
 
-case class Track(path: String, title: String, album: Album, artist: String, duration: Long) 
+case class Track(path: String, title: String, album: Album, artist: String, duration: Int) 
 
 case class Album(title: String, coverArt: Drawable)
